service_desc/requirements.txt
Django==2.2
django-crispy-forms==1.8.1
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
